Implementation of GameManager
It get's some of the load off of the main class and the main frame class. It makes for much less messy code and is generally easier to work with. I also technically did most of the work for the save and load system, so that should basically work now.